+++ cli/cli/kurtosis_config/kurtosis_config_initializer.go
@@ -2,7 +2,6 @@ package kurtosis_config
 
 import (
 	"fmt"
-	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/metrics_optin"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/prompt_displayer"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/user_consent_to_send_metrics_election"
 	"github.com/kurtosis-tech/stacktrace"
@@ -11,15 +10,14 @@ import (
 
 const (
 	metricsConsentPromptLabel       = "Is it okay to send anonymized metrics purely to improve the product?"
-	secondMetricsConsentPromptLabel = "That's okay; we understand. Would it be alright if we send a one-time event recording your opt-out so we can see how much users dislike the metrics? Regardless of your choice, no other events will be tracked per your election."
+	secondMetricsConsentPromptLabel = "Is it alright if we record your opt-out as a one-time event so we can see how much users dislike metrics?"
 
 	shouldSendMetricsDefaultValue = true
 	shouldSendMetricsOptOutEventDefaultValue = true
 )
 
 func initInteractiveConfig() (*KurtosisConfig, error) {
-
-	fmt.Println(metrics_optin.WhyKurtosisCollectMetricsDescriptionNote)
+	printMetricsPreface()
 
 	didUserAcceptSendingMetrics, err := prompt_displayer.DisplayConfirmationPromptAndGetBooleanResult(metricsConsentPromptLabel, shouldSendMetricsDefaultValue)
 	if err != nil {
@@ -28,6 +26,7 @@ func initInteractiveConfig() (*KurtosisConfig, error) {
 	didUserConsentToSendMetricsElectionEvent := didUserAcceptSendingMetrics
 
 	if !didUserAcceptSendingMetrics {
+		fmt.Println("That's okay; we understand. No product analytic metrics will be collected from this point forward.")
 		didUserConsentToSendMetricsElectionEvent, err = prompt_displayer.DisplayConfirmationPromptAndGetBooleanResult(secondMetricsConsentPromptLabel, shouldSendMetricsOptOutEventDefaultValue)
 		if err != nil {
 			return nil, stacktrace.Propagate(err, "An error occurred displaying user metrics consent prompt")
@@ -44,4 +43,33 @@ func initInteractiveConfig() (*KurtosisConfig, error) {
 
 	kurtosisConfig := NewKurtosisConfig(didUserAcceptSendingMetrics)
 	return kurtosisConfig, nil
+}
+
+func printMetricsPreface() {
+	fmt.Println("")
+	fmt.Println("============================================================================================")
+	fmt.Println("                               Metrics Election Preface")
+	fmt.Println("============================================================================================")
+	fmt.Println("The Kurtosis CLI has the potential collect anonymized product analytics metrics, and I want")
+	fmt.Println("to explain why this big message is popping up in your face right now.")
+	fmt.Println("")
+	fmt.Println("User metrics are useful for detecting product bugs and seeing feature usage. However, they're")
+	fmt.Println("also heavily abused to invade our privacy. I hate it as much as I'm guessing you do.")
+	fmt.Println("")
+	fmt.Println("Kurtosis is a small startup so product analytics metrics are vital for us to determine")
+	fmt.Println("where to put resources, but it was important to me that we do this ethically.")
+	fmt.Println("")
+	fmt.Println("Therefore, we've made our metrics:")
+	fmt.Println(" - private: we will *never* give or sell your data to third parties")
+	fmt.Println(" - opt-in: we require you to make a choice about collection rather than assuming you want in")
+	fmt.Println(" - anonymized: your user ID is a hash; we don't know who you are")
+	fmt.Println(" - obfuscated: potentially-sensitive parameters (like module exec params) are hashed as well")
+	fmt.Println("")
+	fmt.Println("If that sounds fair to you, we'd really appreciate you helping us get the data to make our product")
+	fmt.Println("better. In exchange, you have my personal word to honor the trust you place in us by fulfilling")
+	fmt.Println("our metrics promises above.")
+	fmt.Println("")
+	fmt.Println("Sincerely,")
+	fmt.Println("Kevin Today, Kurtosis CTO")
+	fmt.Println("")
 }