+++ cli/docs/changelog.md
@@ -8,6 +8,17 @@
 ### Fixes
 * Fix `network_soft_partition_test` in Golang and Typescript internal testsuite
 
+# 0.9.2
+### Fixes
+* Fixed a bug with `engine restart` using `stacktrace.Propagate` incorrectly
+* The yes/no prompt now displays valid options
+
+### Removals
+* On further review, removed the big long metrics paragraph from the `config init` helptext; this likely isn't going to land well
+
+### Changes
+* Be more explicit in our explanation to users about product analytics metrics
+
 # 0.9.1
 ### Features
 * The enclave ID argument to `enclave inspect` is now tab-completable
@@ -32,7 +43,8 @@
 * Track if user consent sending metrics to improve the product
 
 ### Breaking Changes
-* Upgraded the engine server to 1.9.2
+* Required the user to make an election about whether to send product analytic metrics
+  * **Users using Kurtosis in CI will need to initialize the configuration as part of their CI steps [using these instructions](https://docs.kurtosistech.com/running-in-ci.html)**
   * Users will need to run `kurtosis engine restart` after upgrading to this version of the CLI
   * Engine API users (e.g. in tests) will need to update to `kurtosis-engine-api-lib` 1.9.0